+++ server/src/main/java/io/druid/client/cache/HybridCache.java
@@ -131,7 +131,7 @@ public class HybridCache implements Cache
   }
 
   @Override
-  public void setSegmentsPopularitiesSnapshot(Map<String, Float> segmentsPopularitiesSnapshot)
+  public void setSegmentsPopularitiesSnapshot(Map<String, Double> segmentsPopularitiesSnapshot)
   {
 
   }
