+++ server/src/main/java/io/druid/client/cache/MapCache.java
@@ -177,7 +177,7 @@ public class MapCache implements Cache
   }
 
   @Override
-  public void setSegmentsPopularitiesSnapshot(Map<String, Float> segmentsPopularitiesSnapshot)
+  public void setSegmentsPopularitiesSnapshot(Map<String, Double> segmentsPopularitiesSnapshot)
   {
 
   }
