+++ server/src/main/java/io/druid/client/cache/MemcachedCache.java
@@ -625,7 +625,7 @@ public class MemcachedCache implements Cache
   }
 
   @Override
-  public void setSegmentsPopularitiesSnapshot(Map<String, Float> segmentsPopularitiesSnapshot)
+  public void setSegmentsPopularitiesSnapshot(Map<String, Double> segmentsPopularitiesSnapshot)
   {
 
   }
